nwchem: Fix basis set coordinates selection
Gaussian basis sets can be defined using either spherical or Cartesian
coordinates, and to match the expected energy values calculations using
them should be performed using the same coordinate systems. Unfortunately
NWChem defaults to Cartesian coordinates for all basis sets, and we have
to manually specify which type to use.

A more correct way of doing this would be to inspect the basis set file
itself at runtime, but for now we just hard code a list of the Cartesian
basis sets included in NWChem 7.0.2.